Implement day 19
Part 1 was a pretty easy naive implementation.  I got stuck on part 2
for a while, so I went to sleep.  After sleeping on it, I realized I
could simply keep all possibilities during the expansion because the
self referencing rules all started with a non-self reference as their
first option.