Implement day 20, part 1
Blech, I hate dealing with rotations and stuff.  Then I realized I could
super cheat to find the corners by just counting the number of matching
things.  So I did that.  Now it wants me to actually search the image
for stuff and that is a pain in the butt, so I am stopping for now.