Implement day23
Start with a nice and rust style implementation... it worked but wasn't
fast enough for part2.  Discover that linked lists and rust do not play
nicely together.  Bust out a ton of unsafe and rock on.  Realize I need
a hashmap to make finding the next element efficient too.  Spend
entirely too long looking for an off by one error.